devtools/client/shared/components/reps/test/mochitest/test_reps_stylesheet.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test Stylesheet rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - Stylesheet</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  try {
    const {
      REPS,
      getRep,
    } = browserRequire("devtools/client/shared/components/reps/reps");
    let { StyleSheet } = REPS;

    let gripStub = {
      "type": "object",
      "class": "CSSStyleSheet",
      "actor": "server1.conn2.obj1067",
      "extensible": true,
      "frozen": false,
      "sealed": false,
      "ownPropertyLength": 0,
      "preview": {
        "kind": "ObjectWithURL",
        "url": "https://example.com/styles.css"
      }
    };

    // Test that correct rep is chosen
    is(getRep(gripStub), StyleSheet.rep, "Rep correctly selects StyleSheet Rep");

    // Test rendering
    const renderedComponent = renderComponent(StyleSheet.rep, { object: gripStub });
    is(renderedComponent.textContent, "StyleSheet https://example.com/styles.css", "StyleSheet rep has expected text content");

    // Test rendering with objectLink
    const objectLinkRenderedComponent = renderComponent(StyleSheet.rep, {
      object: gripStub,
      objectLink: (props, ...children) => React.DOM.span({},
        "*", ...children, "*"),
    });
    is(
      objectLinkRenderedComponent.textContent,
      "*StyleSheet *https://example.com/styles.css",
      "StyleSheet rep has expected text content when an objectLink is passed as a prop"
    );
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
